requirements.txt
altgraph==0.17.2
arrow==1.2.2
asgiref==3.5.0
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
comtypes==1.1.11
constantly==15.1.0
cookiecutter==1.7.3
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-browser-reload==1.3.0
django-ckeditor==6.3.2
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-tailwind==3.1.1
django-tinymce==3.4.0
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.6.0
Flask==2.0.2
future==0.18.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
jinja2-time==0.2.0
Js2Py==0.71
MarkupSafe==2.0.1
MouseInfo==0.1.3
oauthlib==3.2.0
packaging==21.3
pefile==2021.9.3
Pillow==9.1.0
pipwin==0.5.2
platformdirs==2.5.1
poyo==0.5.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydub==0.25.1
pygame==2.1.2
PyGetWindow==0.0.9
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pyjokes==0.6.0
pyjsparser==2.7.1
PyJWT==2.3.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-slugify==6.1.1
python3-openid==3.2.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pywin32==303
pywin32-ctypes==0.2.0
razorpay==1.3.0
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.2
urllib3==1.26.8
virtualenv==20.13.4
Werkzeug==2.0.3
whitenoise==6.0.0
wikipedia==1.4.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zope.interface==5.4.0
